### 2025-06-09 #8374 @editorialbot
[PRE REVIEW]: KrakenParser: Efficient Post-processing of Kraken2-like Taxonomic Reports
**Submitting author:** @iliapopov17 (<a href="http://orcid.org/0000-0001-7947-1654">Ilia Popov</a>)
**Repository:** https://github.com/PopovIILab/KrakenParser
**Branch with paper.md** (empty if default branch): dev
**Version:** v0.5.0
