### 2025-06-23 6a849d2
add diversity function
