### 2025-06-23 aeca828
add diversity function
