### 2025-09-25 #8374 @fboehm
> [@editorialbot](https://github.com/editorialbot) invite [@fboehm](https://github.com/fboehm) as editor
>
> Fred, might you be able to handle this one? It's been waiting for some time, and it seems to fit your area reasonably!

Sorry to have overlooked this, @rkurchin. I'm unable to take on another review right now.
