### 2026-03-12 8367c2e
build: add pyproject.toml and migrate packaging from wheel metadata
