### 2026-03-16 3ac8555
refactor(mpa): rewrite kreport2mpa as transform2mpa with batch mode
Add -i/--input dir mode alongside the existing single-file -r mode.
Remove run_kreport2mpa.sh wrapper — batch logic now lives in Python.
