### 2026-03-24 3ccdab8
refactor(counts): replace bash decombine scripts with Python split_mpa
Merge decombine.sh + decombine_viruses.sh into a single split_mpa.py.
Add --viruses-only and --keep-human flags.
Move convert2csv and processing_script into counts/ subpackage.
