+++ krakenparser/counts/processing_script.py
@@ -1,7 +1,8 @@
 #!/usr/bin/env python
 
+import os
 import sys
-import shutil
+import tempfile
 import argparse
 from pathlib import Path
 
@@ -10,40 +11,44 @@ def modify_taxa_names(line):
     prefixes = ["s__", "g__", "f__", "o__", "c__", "p__"]
     for prefix in prefixes:
         if line.startswith(prefix):
-            # Remove the prefix and replace underscores with spaces
-            return line[len(prefix) :].replace("_", " ")
+            parts = line[len(prefix):].split("\t")
+            parts[0] = parts[0].replace("_", " ")
+            return "\t".join(parts)
     return line
 
 
 def process_files(source_file, destination_file):
+    src_path = Path(source_file)
+    if not src_path.is_file():
+        raise FileNotFoundError(f"Source file not found: {src_path}")
+    dest_path = Path(destination_file)
+    if not dest_path.is_file():
+        raise FileNotFoundError(f"Destination file not found: {dest_path}")
+
     # Read the first line from the source file and modify it
-    with open(source_file, "r") as file:
+    with open(src_path, "r") as file:
         first_line_source = file.readline()
     modified_first_line = "\t".join(
         word.split(".")[0] for word in first_line_source.split()
     )
 
     # Read all content from the destination file and modify taxa names
-    with open(destination_file, "r") as file:
+    with open(dest_path, "r") as file:
         lines = file.readlines()
     modified_lines = [modify_taxa_names(line.strip()) for line in lines]
 
     # Combine the modified first line with the modified content of the destination file
     updated_content = modified_first_line + "\n" + "\n".join(modified_lines)
 
-    # Write the updated content back to the destination file
-    with open(destination_file, "w") as file:
-        file.write(updated_content)
+    # Write atomically: write to a temp file in the same directory, then replace
+    with tempfile.NamedTemporaryFile(
+        mode="w", dir=dest_path.parent, delete=False, suffix=".tmp"
+    ) as tmp:
+        tmp.write(updated_content)
+        tmp_path = tmp.name
+    os.replace(tmp_path, dest_path)
 
     print(f"Processed {destination_file} successfully.")
-
-    # Get the path to the current directory (same location as the script)
-    current_dir = Path(__file__).resolve().parent
-    pycache_dir = current_dir / "__pycache__"
-
-    # Check if __pycache__ exists and remove it
-    if pycache_dir.exists() and pycache_dir.is_dir():
-        shutil.rmtree(pycache_dir)
 
 
 if __name__ == "__main__":
