### 2026-04-01 eb92428
refactor(stats): move relabund and diversity into stats/ subpackage
