### 2026-04-05 58c1d8a
feat(pipeline): port full pipeline orchestration from bash to Python
Replace kraken2csv.sh with pipeline.py. Add run_pipeline() as a
callable library function. Add -o/--output and --overwrite flags.
Skip binary and hidden files automatically (_is_processable check).
