+++ krakenparser/pipeline.py
@@ -1,0 +1,159 @@
+#!/usr/bin/env python
+"""Full KrakenParser pipeline: kreport → MPA → combined → counts → rel_abund → diversity."""
+
+import argparse
+import logging
+import shutil
+import sys
+from pathlib import Path
+
+_log = logging.getLogger(__name__)
+
+import pandas as pd
+
+from krakenparser.mpa.transform2mpa import kreport_to_mpa
+from krakenparser.mpa.mpa_table import combine_mpa
+from krakenparser.counts.split_mpa import split_mpa
+from krakenparser.counts.processing_script import process_files
+from krakenparser.counts.convert2csv import convert_to_csv
+from krakenparser.stats.relabund import calculate_rel_abund
+from krakenparser.stats.diversity import calc_alpha_div, calc_beta_div
+
+
+def _is_processable(path: Path) -> bool:
+    """Return False for hidden files, files with null bytes, or non-UTF-8 files."""
+    if path.name.startswith("."):
+        return False
+    try:
+        chunk = path.read_bytes()[:1024]
+        if b"\x00" in chunk:
+            return False
+        chunk.decode("utf-8")
+        return True
+    except (UnicodeDecodeError, OSError):
+        return False
+
+
+_OUTPUT_SUBDIRS = ("intermediate", "counts", "rel_abund", "diversity")
+
+
+def run_pipeline(
+    input_dir: str,
+    output_dir: str | None = None,
+    keep_human: bool = False,
+    rarefaction_depth: int = 1000,
+    overwrite: bool = False,
+) -> None:
+    source_dir = Path(input_dir)
+    if not source_dir.is_dir():
+        sys.exit(f"Error: input directory not found: {source_dir}")
+
+    out_dir = Path(output_dir) if output_dir else source_dir.parent
+    out_dir.mkdir(parents=True, exist_ok=True)
+
+    existing = [out_dir / d for d in _OUTPUT_SUBDIRS if (out_dir / d).exists()]
+    if existing and not overwrite:
+        names = ", ".join(d.name for d in existing)
+        sys.exit(
+            f"Error: output already exists in '{out_dir}' ({names}).\n"
+            "Use --overwrite to overwrite it."
+        )
+    if overwrite:
+        for d in existing:
+            shutil.rmtree(d)
+            _log.info("Removed existing directory: %s", d)
+
+    intermediate_dir = out_dir / "intermediate"
+    intermediate_dir.mkdir(exist_ok=True)
+
+    # Part 1: kreport → MPA
+    mpa_dir = intermediate_dir / "mpa"
+    mpa_dir.mkdir(exist_ok=True)
+    for f in sorted(source_dir.iterdir()):
+        if not f.is_file():
+            continue
+        if not _is_processable(f):
+            _log.info("Skipping: %s", f.name)
+            continue
+        out_name = f.stem + ".MPA.TXT"
+        kreport_to_mpa(str(f), str(mpa_dir / out_name), display_header=True)
+
+    # Part 2: combine MPAs
+    mpa_files = sorted(mpa_dir.glob("*.MPA.TXT"))
+    if not mpa_files:
+        sys.exit("Error: no MPA files found after conversion.")
+    combined_file = intermediate_dir / "COMBINED.txt"
+    combine_mpa([str(f) for f in mpa_files], str(combined_file))
+    _log.info("MPA files combined. Output: %s", combined_file)
+
+    # Part 3: split combined MPA by rank
+    split_mpa(str(combined_file), str(intermediate_dir), keep_human=keep_human)
+    txt_dir = intermediate_dir / "txt"
+
+    # Part 4: clean taxa names and add sample header
+    for txt_file in sorted(txt_dir.glob("counts_*.txt")):
[…]
