### 2026-04-10 2104768
feat(kpplot): add aggregate_by_metadata, fix grid axis, add cmap validation
- Add aggregate_by_metadata() to base for metadata grouping
- Fix streamgraph grid axis (x→y)
- Raise ValueError when cmap list is shorter than taxa count
- Add missing-sample validation on sample_order
