### 2026-04-22 7dd6914
ci: rename workflow to python-package.yml, expand matrix to 3.10–3.14
Add fail-fast: false, flake8 lint step, --no-build-isolation,
branch-scoped triggers (dev, main), codecov-action@v5.
