+++ .github/workflows/ci.yml
@@ -1,23 +1,0 @@
-name: pytest
-on:
-  push:
-    branches: [ dev ]
-  pull_request:
-    branches: [ dev ]
-
-jobs:
-  build-test:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.12]
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-      - run: |
-          python -m pip install --upgrade pip
-          pip install krakenparser
-          pip install pytest
-      - run: pytest -q
