### 2026-05-06 00efa56
test: add kpplot tests and raise coverage to 83%
Add smoke tests for stacked_barplot, streamgraph, clustermap.
Switch test_full_pipeline to direct run_pipeline() call so pipeline.py
is included in coverage. Add overwrite protection tests.
