### 2026-05-12 0aa7b8d
docs: update README for v1.0.0
Add codecov badge. Document new -o/--output and --overwrite flags.
Add Before Visualization section highlighting --relabund -O.
Move step-by-step modules to collapsible <details> block.
Update Example Output Structure with intermediate/ directory.
