+++ README.md
@@ -2,7 +2,8 @@
 
 ![License](https://img.shields.io/badge/License-MIT-steelblue)
 [![Downloads](https://static.pepy.tech/badge/krakenparser)](https://pepy.tech/project/krakenparser)
-[![CI](https://github.com/PopovIILab/KrakenParser/actions/workflows/ci.yml/badge.svg)](https://github.com/PopovIILab/KrakenParser/actions/workflows/ci.yml)
+[![CI](https://github.com/PopovIILab/KrakenParser/actions/workflows/python-package.yml/badge.svg)](https://github.com/PopovIILab/KrakenParser/actions/workflows/python-package.yml)
+[![codecov](https://codecov.io/gh/PopovIILab/KrakenParser/graph/badge.svg)](https://codecov.io/gh/PopovIILab/KrakenParser)
 
 <img src="https://github.com/PopovIILab/KrakenParser/blob/main/imgs/KrakenParser_logo_light.png#gh-light-mode-only" align="left"/>
 <img src="https://github.com/PopovIILab/KrakenParser/blob/main/imgs/KrakenParser_logo_dark.png#gh-dark-mode-only" align="left"/>
@@ -108,9 +109,10 @@ X9,0.7232472324723247,0.7352941176470589,...,0.8066914498141264,0.0
 ## Quick Start (Full Pipeline)
 To run the full pipeline, use the following command:
 ```bash
-KrakenParser --complete -i data/kreports
+KrakenParser --complete -i data/kreports -o results/
 #Having troubles? Run KrakenParser --complete -h
 ```
+
 This will:
 1. Convert Kraken2 reports to MPA format
 2. Combine MPA files into a single file
@@ -120,162 +122,176 @@ This will:
 6. Calculate relative abundance
 7. Calculate α & β-diversities
 
-### **Input Requirements**
-- The Kraken2 reports must be inside a **subdirectory** (e.g., `data/kreports`).
-- The script automatically creates output directories and processes the data.
-
 ## Installation
 
 ```
 pip install krakenparser
 ```
 
-## Using Individual Modules
-You can also run each step manually if needed.
+## Before Visualization: Grouping Low-Abundance Taxa
+
+The full pipeline automatically calculates relative abundance. Before passing data to visualization, it is strongly recommended to re-run `--relabund` with the `-O` flag — this collapses all taxa below the chosen threshold into a single **"Other"** group, producing much cleaner and more readable plots.
+
+```bash
+KrakenParser --relabund -i data/counts/counts_species.csv -o data/rel_abund/ra_species.csv -O 4
+```
+
+This groups every taxon with relative abundance **< 4 %** into `Other (<4.0%)`. Adjust the threshold to your data.
+
+> **Note:** The pipeline-generated `rel_abund/ra_*.csv` files (no `-O`) preserve the full unfiltered data — use them for statistical analysis. Use the `-O` variant specifically for visualization.
+
+---
+
+<details>
+<summary><b>Using Individual Modules (Advanced)</b></summary>
+<br>
+
+Each step of the pipeline can also be run individually. This is useful for re-running a single step, debugging, or integrating KrakenParser into a custom workflow.
 
 ### **Step 1: Convert Kraken2 Reports to MPA Format**
 ```bash
-KrakenParser --kreport2mpa -i data/kreports -o data/mpa
+# Batch mode (directory)
+KrakenParser --kreport2mpa -i data/kreports -o data/intermediate/mpa
+# Single file
[…]
