### 2026-05-12 9664a56
ci: pre-install numpy to fix biom-format build on Python 3.14
biom-format uses a legacy setup.py that imports numpy at metadata
generation time. Pre-installing numpy ensures it is available before
pip resolves scikit-bio's transitive dependencies.
