+++ .github/workflows/python-package.yml
@@ -27,7 +27,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install flake8 setuptools wheel
+        python -m pip install flake8 setuptools wheel numpy
         pip install -e ".[dev]" --no-build-isolation
     - name: Lint with flake8
       run: |
