### 2026-05-12 8d5d457
ci: drop Python 3.14 from matrix; cap requires-python to <=3.13
biom-format (transitive dep of scikit-bio) has no cp314 wheel and
its legacy setup.py has undeclared build-time deps (numpy, Cython).
Pin support to 3.10–3.13 until scikit-bio ecosystem catches up.
