### 2026-05-12 06ddafb
refactor(stats): replace scikit-bio with scipy for beta diversity
scikit-bio pulls in biom-format which has no cp314 wheel and
brittle build-time deps. Reimplement subsample_counts in numpy
and use scipy.spatial.distance for Bray-Curtis and Jaccard.
Restore Python 3.14 support and bump requires-python to <=3.16.
