+++ pyproject.toml
@@ -9,13 +9,13 @@ description = "A collection of scripts designed to process Kraken2 reports and c
 readme = {file = "README.md", content-type = "text/markdown"}
 license = {file = "LICENSE"}
 authors = [{name = "Ilia Popov", email = "iljapopov17@gmail.com"}]
-requires-python = ">=3.10,<=3.13"
+requires-python = ">=3.10,<=3.16"
 dependencies = [
     "pandas>=2.0.0,<=2.3.3",
     "matplotlib>=3.7.0,<=3.10.8",
     "numpy>=1.24.0,<=2.3.5",
     "seaborn>=0.12.0,<=0.13.2",
-    "scikit-bio>=0.6.0,<=0.7.2",
+    "scipy>=1.9.0,<=1.16.3",
 ]
 
 [project.optional-dependencies]
