### 2026-05-12 4ce8750
feat(diversity): add -s/--seed for reproducible rarefaction
Rarefaction is stochastic — without a fixed seed results differ
between runs. Replace np.random.choice with np.random.default_rng
and propagate the seed through calc_beta_div and run_pipeline.
