### 2026-05-12 185e8d4
docs: document -s/--seed flag in Quick Start and diversity step
