### 2026-05-12 447a414
build: exclude paper/ and .github/ from sdist
