+++ MANIFEST.in
@@ -1,0 +1,2 @@
+prune paper
+prune .github
