### 2026-05-18 752d2d9
style(scripts): apply Ruff formatter and import sorting
