+++ krakenparser/counts/convert2csv.py
@@ -3,6 +3,7 @@
 import argparse
 import logging
 from pathlib import Path
+
 import pandas as pd
 
 _log = logging.getLogger(__name__)
