### 2026-05-18 34616e5
ci: replace flake8 with Ruff linter and formatter
