### 2026-05-18 3f6f0b4
ci: add codecov.yml with informational patch coverage
