+++ codecov.yml
@@ -1,0 +1,6 @@
+# codecov.yml
+coverage:
+  patch:
+    target: 78%   # снизить порог
+    # или:
+    informational: true  # не фейлить, только информировать
