### 2026-05-18 bc0f1e5
style(ci): remove comments from workflow file
