### 2026-05-18 4dc34dc
revert: replace flake8 with Ruff linter and formatter
