### 2026-05-22 15267a4
refactor(types): fix Pylance warnings in calc_beta_div
