### 2026-05-22 26afe67
feat(utils): add ensure_output_dir helper
