### 2026-05-22 3993806
refactor(mpa): use ensure_output_dir, add logging and input validation
