### 2026-05-22 c320f0f
refactor(stats): use ensure_output_dir, add logging and startup info
