### 2026-05-22 17393de
refactor(pipeline): raise exceptions instead of sys.exit in run_pipeline
