### 2026-05-22 2f51ffb
test: add ensure_output_dir unit tests
