### 2026-05-22 758b07e
test: cover auto-create dirs, new validations, new exception types
