### 2026-05-22 87b9da0
test(pipeline): fix overwrite test — FileExistsError, not SystemExit
