### 2026-05-22 fff6234
test: add CLI smoke tests for all main() entry points
