+++ tests/test_cli.py
@@ -1,0 +1,199 @@
+"""Smoke tests for CLI entry-points (main() functions via sys.argv monkeypatching)."""
+
+import shutil
+import sys
+import warnings
+
+import pandas as pd
+import pytest
+
+from krakenparser.counts.convert2csv import main as convert2csv_main
+from krakenparser.counts.processing_script import main as processing_main
+from krakenparser.counts.split_mpa import main as split_mpa_main
+from krakenparser.mpa.mpa_table import main as mpa_table_main
+from krakenparser.mpa.transform2mpa import main as transform2mpa_main
+from krakenparser.pipeline import main as pipeline_main
+from krakenparser.stats.diversity import main as diversity_main
+from krakenparser.stats.relabund import main as relabund_main
+
+_MPA_A = "#Classification\tsample1\nd__Bacteria|s__Pseudomonas_aeruginosa\t300\n"
+_MPA_B = "#Classification\tsample2\nd__Bacteria|s__Pseudomonas_aeruginosa\t100\n"
+
+_COMBINED_MPA = (
+    "#Classification\tsample1\tsample2\n"
+    "d__Bacteria|p__Pseudomonadota|g__Pseudomonas|s__Pseudomonas_aeruginosa\t300\t100\n"
+    "d__Bacteria|p__Bacteroidota\t100\t80\n"
+)
+
+
+# ---------------------------------------------------------------------------
+# convert2csv
+# ---------------------------------------------------------------------------
+
+
+def test_convert2csv_main(counts_txt_file, tmp_path, monkeypatch):
+    out = tmp_path / "out.csv"
+    monkeypatch.setattr(sys, "argv", ["c2c", "-i", str(counts_txt_file), "-o", str(out)])
+    convert2csv_main()
+    assert out.exists()
+
+
+# ---------------------------------------------------------------------------
+# processing_script
+# ---------------------------------------------------------------------------
+
+
+def test_processing_main(tmp_path, monkeypatch):
+    source = tmp_path / "COMBINED.txt"
+    source.write_text("#Classification\tsample1.kreport\n")
+    dest = tmp_path / "counts.txt"
+    dest.write_text("s__Pseudomonas_aeruginosa\t100\n")
+    monkeypatch.setattr(sys, "argv", ["ps", "-i", str(source), "-o", str(dest)])
+    processing_main()
+
+
+# ---------------------------------------------------------------------------
+# split_mpa
+# ---------------------------------------------------------------------------
+
+
+def test_split_mpa_main(tmp_path, monkeypatch):
+    combined = tmp_path / "COMBINED.txt"
+    combined.write_text(_COMBINED_MPA)
+    out = tmp_path / "out"
+    monkeypatch.setattr(sys, "argv", ["sm", "-i", str(combined), "-o", str(out)])
+    split_mpa_main()
+    assert (out / "txt" / "counts_species.txt").exists()
+
+
+def test_split_mpa_main_viruses_only(tmp_path, monkeypatch):
+    combined = tmp_path / "COMBINED.txt"
+    combined.write_text(_COMBINED_MPA + "d__Viruses|s__Virus_X\t5\t3\n")
+    out = tmp_path / "out"
+    monkeypatch.setattr(
+        sys, "argv", ["sm", "-i", str(combined), "-o", str(out), "--viruses-only"]
+    )
+    split_mpa_main()
+
+
+def test_split_mpa_main_keep_human(tmp_path, monkeypatch):
+    combined = tmp_path / "COMBINED.txt"
+    combined.write_text(_COMBINED_MPA)
+    out = tmp_path / "out"
+    monkeypatch.setattr(
+        sys, "argv", ["sm", "-i", str(combined), "-o", str(out), "--keep-human"]
[…]
