### 2026-05-22 da16a3b
test: cover edge cases — _strip_path_prefix, _is_processable, t__ filter
