### 2026-05-22 ae1a78c
style(scripts): apply Ruff formatter
