### 2026-05-22 bf0f182
build: exclude __main__ guard lines from coverage reporting
