+++ pyproject.toml
@@ -48,3 +48,9 @@ omit = [
     "krakenparser/mpa/transform2mpa.py",
     "krakenparser/mpa/mpa_table.py",
 ]
+
+[tool.coverage.report]
+exclude_lines = [
+    "pragma: no cover",
+    "if __name__ == .__main__.:",
+]
