### 2026-05-31 3cedb39
refactor(docs): update CLI usage to use --step flag
