+++ README.md
@@ -107,15 +107,9 @@ X9,0.7232472324723247,0.7352941176470589,...,0.8066914498141264,0.0
 |![combined_white](https://github.com/user-attachments/assets/48b3f6e3-6dd5-4298-a793-23dcd549e90c)|![kpclust](https://github.com/user-attachments/assets/98a4d540-7c43-4802-8f77-277a5637a7a1)|
 
 ## Quick Start (Full Pipeline)
-To run the full pipeline, use the following command:
-```bash
-KrakenParser --complete -i data/kreports -o results/
-#Having troubles? Run KrakenParser --complete -h
-```
 
-For **reproducible** β-diversity (rarefaction is stochastic by default):
 ```bash
-KrakenParser -i data/kreports -o results/ -s 42
+KrakenParser -i data/kreports -o results/
 ```
 
 This will:
@@ -127,23 +121,46 @@ This will:
 6. Calculate relative abundance
 7. Calculate α & β-diversities
 
+> [!TIP]
+> After the pipeline finishes, the output window will remind you about calibrating
+> rarefaction depth for β-diversity and re-running relative abundance normalization
+> before visualization — with ready-to-paste example commands tailored to your output paths.
+
+### Full help output
+
+```
+usage: KrakenParser [-h] [-i INPUT] [-o OUTPUT] [--viruses] [--keep-human]
+                    [-V] [-d DEPTH] [-s SEED] [--overwrite]
+                    [--step {mpa,combine,split,process,csv,relabund,diversity}]
+
+KrakenParser: Convert Kraken2 Reports to CSV.
+
+options:
+  -h, --help            show this help message and exit
+
+Core Arguments:
+  -i, --input INPUT     Directory containing Kraken2 report files
+  -o, --output OUTPUT   Output directory (default: parent of input)
+  --viruses             Extract only VIRUSES domain taxa in the pipeline
+  --keep-human          Do not filter human-related taxa
+  -V, --version         show program's version number and exit
+
+Pipeline Options (Full Run):
+  -d, --depth DEPTH     Rarefaction depth for β-diversity (default: 1000)
+  -s, --seed SEED       Random seed for reproducible rarefaction (default: random)
+  --overwrite           Overwrite the output directory if it already exists
+
+Advanced (Step-by-step control):
+  --step {mpa,combine,split,process,csv,relabund,diversity}
+                        Run only a specific part of the pipeline.
+                        Type 'krakenparser --step <name> -h' for more.
+```
+
 ## Installation
 
 ```
 pip install krakenparser
 ```
-
-## Before Visualization: Grouping Low-Abundance Taxa
-
-The full pipeline automatically calculates relative abundance. Before passing data to visualization, it is strongly recommended to re-run `--relabund` with the `-O` flag — this collapses all taxa below the chosen threshold into a single **"Other"** group, producing much cleaner and more readable plots.
-
-```bash
-KrakenParser --relabund -i data/counts/counts_species.csv -o data/rel_abund/ra_species.csv -O 4
-```
-
-This groups every taxon with relative abundance **< 4 %** into `Other (<4.0%)`. Adjust the threshold to your data.
-
-> **Note:** The pipeline-generated `rel_abund/ra_*.csv` files (no `-O`) preserve the full unfiltered data — use them for statistical analysis. Use the `-O` variant specifically for visualization.
 
 ---
 
@@ -151,123 +168,118 @@ This groups every taxon with relative abundance **< 4 %** into `Other (<4.0%)`. 
[…]
