### 2026-05-31 00a381f
refactor: modernize CLI routing and implement --step interface
