### 2026-05-31 e15d1ef
refactor(mpa): implement hierarchical sorting for combined MPA tables
