+++ krakenparser/mpa/mpa_table.py
@@ -12,9 +12,12 @@ _log = logging.getLogger(__name__)
 
 def combine_mpa(in_files: list[str], o_file: str) -> None:
     out_path = ensure_output_dir(o_file, is_file=True)
-    # Plain dict preserves insertion order (Python 3.7+).
-    taxa: dict[str, dict[int, str]] = {}
-    sample_names: list[str] = []
+
+    samples: dict[int, str] = {}
+    values: dict[str, dict[int, str]] = {}
+    parent2child: dict[str, list[str]] = {}
+    toparse: list[str] = []
+    sample_count = 0
 
     _log.info("Number of files to parse: %d", len(in_files))
 
@@ -22,8 +25,10 @@ def combine_mpa(in_files: list[str], o_file: str) -> None:
         if not Path(in_path).is_file():
             raise FileNotFoundError(f"Input file not found: {in_path}")
 
-    for idx, in_path in enumerate(in_files):
-        sample_name = f"Sample #{idx + 1}"
+    for in_path in in_files:
+        sample_count += 1
+        sample_name = f"Sample #{sample_count}"
+
         with open(in_path) as fh:
             for line in fh:
                 line = line.rstrip("\n")
@@ -34,26 +39,56 @@ def combine_mpa(in_files: list[str], o_file: str) -> None:
                     if len(cols) >= 2:
                         sample_name = cols[-1]
                     continue
+
                 cols = line.split("\t", 1)
                 if len(cols) < 2:
                     continue
-                taxon, count = cols[0], cols[1]
-                if taxon not in taxa:
-                    taxa[taxon] = {}
-                taxa[taxon][idx] = count
-        sample_names.append(sample_name)
+                classification, val = cols[0], cols[1]
 
-    n_samples = len(sample_names)
-    n_taxa = len(taxa)
+                split_vals = classification.split("|")
+                curr_parent = ""
+                for i in range(len(split_vals)):
+                    test_val = "|".join(split_vals[:i])  # при i=0 → ""
+                    if test_val in values:
+                        curr_parent = test_val
+
+                if curr_parent == "":
+                    if classification not in toparse:
+                        toparse.append(classification)
+                else:
+                    if curr_parent not in parent2child:
+                        parent2child[curr_parent] = []
+                    if classification not in parent2child[curr_parent]:
+                        parent2child[curr_parent].append(classification)
+
+                if classification not in values:
+                    values[classification] = {}
+                values[classification][sample_count] = val
+
+        samples[sample_count] = sample_name
+
+    n_taxa = len(values)
     _log.info("Number of classifications to write: %d", n_taxa)
 
+    count_written = 0
     with open(out_path, "w") as fh:
-        fh.write("#Classification\t" + "\t".join(sample_names) + "\n")
-        for taxon, counts in taxa.items():
-            row = [counts.get(i, "0") for i in range(n_samples)]
-            fh.write(taxon + "\t" + "\t".join(row) + "\n")
+        header = "#Classification\t" + "\t".join(
+            samples[i] for i in range(1, sample_count + 1)
+        )
+        fh.write(header + "\n")
 
-    _log.info("%d classifications written", n_taxa)
+        stack = list(toparse)
+        while stack:
+            curr = stack.pop(0)
+            if curr in parent2child:
+                stack = parent2child[curr] + stack
+            row = "\t".join(
+                values[curr].get(i, "0") for i in range(1, sample_count + 1)
+            )
+            fh.write(curr + "\t" + row + "\n")
+            count_written += 1
+
+    _log.info("%d classifications written", count_written)
 
 
 def main() -> None:
