### 2026-05-31 e48266e
refactor(mpa): overhaul kreport parsing logic for robust hierarchy tracking
