+++ krakenparser/mpa/transform2mpa.py
@@ -9,51 +9,65 @@ from pathlib import Path
 
 from krakenparser.utils import ensure_output_dir
 
-# Maps Kraken2 single-letter rank codes to MPA prefixes
-_RANK_PREFIX = {
-    "D": "d",
-    "K": "k",
-    "P": "p",
-    "C": "c",
-    "O": "o",
-    "F": "f",
-    "G": "g",
-    "S": "s",
-}
-
 _log = logging.getLogger(__name__)
 
+_MAIN_LVLS = {"R", "K", "D", "P", "C", "O", "F", "G", "S"}
 
-def _parse_line(line: str):
+
+def _parse_line(line: str, remove_spaces: bool = False) -> list:
     """
     Parse one Kraken2 report line.
 
-    Standard format (6 columns):
-        pct  cum_reads  direct_reads  rank  taxid  name(indented)
-
-    Returns (name, depth, rank, cum_reads, pct) or None on malformed input.
+    Returns [name, level_num, level_type, all_reads, percents]
+    or empty list on malformed input.
     """
     parts = line.rstrip("\n").split("\t")
-    if len(parts) < 5:
-        return None
+    if len(parts) < 4:
+        return []
     try:
-        pct = float(parts[0])
-        cum_reads = int(parts[1])
+        int(parts[1])
     except ValueError:
-        return None
+        return []
 
-    rank = parts[3].strip()
-    name_field = parts[-1]  # always the last column regardless of format variant
+    try:
+        percents = float(parts[0])
+    except ValueError:
+        return []
+    all_reads = int(parts[1])
 
-    depth = 0
-    for ch in name_field:
+    try:
+        int(parts[-3])
+        level_type = parts[-2].strip()
+        map_kuniq = {
+            "species": "S",
+            "genus": "G",
+            "family": "F",
+            "order": "O",
+            "class": "C",
+            "phylum": "P",
+            "superkingdom": "D",
+            "kingdom": "K",
+        }
+        if level_type not in map_kuniq:
+            level_type = "-"
+        else:
+            level_type = map_kuniq[level_type]
+    except ValueError:
+        level_type = parts[-3].strip()
+
+    name = parts[-1]
+    spaces = 0
+    for ch in name:
         if ch == " ":
-            depth += 1
+            spaces += 1
         else:
             break
-    name = name_field.strip()
+    name = name.strip()
+    if remove_spaces:
+        name = name.replace(" ", "_")
 
-    return name, depth // 2, rank, cum_reads, pct
+    level_num = spaces / 2
+    return [name, level_num, level_type, all_reads, percents]
 
 
 def kreport_to_mpa(
@@ -67,54 +81,63 @@ def kreport_to_mpa(
     """
     Convert a single Kraken2 report to MPA format.
 
-    Uses a stack to track the current taxonomic path. Each entry is
-    (structural_depth, mpa_segment, is_standard_rank). When a node at
-    depth d is encountered, all stack entries with depth >= d are popped
-    before the new entry is pushed, keeping the path consistent.
+    Tracks the current taxonomic path via curr_path and prev_lvl_num,
+    popping the stack when moving back up the tree — exactly as the
+    original script does.
     """
     if not Path(report_path).is_file():
         raise FileNotFoundError(f"Input file not found: {report_path}")
     out_path = ensure_output_dir(output_path, is_file=True)
-    # Stack entries: (structural_depth, mpa_segment, is_standard_rank)
-    stack: list[tuple[int, str, bool]] = []
+
+    curr_path: list[str] = []
+    prev_lvl_num = -1
 
     with open(report_path) as r_fh, open(out_path, "w") as o_fh:
         if display_header:
[…]
