### 2026-05-31 54bf8e7
refactor: improve CLI startup output visibility
