### 2026-05-31 766ce52
docs: update results directory structure in README
