### 2026-06-01 234c7d0
feat(mpa): migrate CLI from argparse to typer
