+++ krakenparser/mpa/mpa_table.py
@@ -1,17 +1,26 @@
 #!/usr/bin/env python
 """Combine multiple MPA-format files into a single merged table."""
 
-import argparse
 import logging
+import sys
 from pathlib import Path
+from typing import Optional
+
+import typer
 
 from krakenparser.utils import ensure_output_dir
 
 _log = logging.getLogger(__name__)
 
+app = typer.Typer(
+    name="combine",
+    add_completion=False,
+    context_settings={"help_option_names": ["-h", "--help"]},
+)
 
-def combine_mpa(in_files: list[str], o_file: str) -> None:
-    out_path = ensure_output_dir(o_file, is_file=True)
+
+def combine_mpa(in_files: list[Path], o_file: Path) -> None:
+    out_path = ensure_output_dir(str(o_file), is_file=True)
 
     samples: dict[int, str] = {}
     values: dict[str, dict[int, str]] = {}
@@ -22,7 +31,7 @@ def combine_mpa(in_files: list[str], o_file: str) -> None:
     _log.info("Number of files to parse: %d", len(in_files))
 
     for in_path in in_files:
-        if not Path(in_path).is_file():
+        if not in_path.is_file():
             raise FileNotFoundError(f"Input file not found: {in_path}")
 
     for in_path in in_files:
@@ -91,29 +100,42 @@ def combine_mpa(in_files: list[str], o_file: str) -> None:
     _log.info("%d classifications written", count_written)
 
 
-def main() -> None:
-    logging.basicConfig(level=logging.INFO, format="%(message)s")
-    parser = argparse.ArgumentParser(
-        description="Combine MPA files into a single tab-delimited table."
-    )
-    parser.add_argument(
+@app.callback(invoke_without_command=True)
+def main(
+    ctx: typer.Context,  # Контекст для работы с хелпом
+    in_files: Optional[list[Path]] = typer.Option(
+        None,
         "-i",
         "--input",
-        required=True,
-        nargs="+",
-        dest="in_files",
-        help="Input MPA files (one per sample)",
-    )
-    parser.add_argument(
+        help="Input MPA files (one per sample). Повторите флаг -i для каждого файла.",
+    ),
+    o_file: Optional[Path] = typer.Option(
+        None,
         "-o",
         "--output",
-        required=True,
-        dest="o_file",
-        help="Output merged MPA file",
-    )
-    args = parser.parse_args()
-    combine_mpa(args.in_files, args.o_file)
+        help="Output merged MPA file.",
+    ),
+) -> None:
+    """Combine MPA files into a single tab-delimited table."""
+    logging.basicConfig(level=logging.INFO, format="%(message)s")
+
+    if not in_files and o_file is None:
+        print(ctx.get_help())
+        raise typer.Exit()
+
+    if not in_files or o_file is None:
+        print(
+            "Error: Missing required options '-i / --input' and '-o / --output'.",
+            file=sys.stderr,
+        )
+        raise typer.Exit(code=1)
+
+    try:
+        combine_mpa(in_files, o_file)
+    except FileNotFoundError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        raise typer.Exit(code=1)
 
 
 if __name__ == "__main__":
-    main()
+    app()
