### 2026-06-01 abbda67
feat(counts): migrate convert2csv CLI from argparse to typer
