+++ krakenparser/counts/convert2csv.py
@@ -1,17 +1,24 @@
 #!/usr/bin/env python
-
-import argparse
 import logging
+import sys
 from pathlib import Path
+from typing import Optional
 
 import pandas as pd
+import typer
 
 from krakenparser.utils import ensure_output_dir
 
 _log = logging.getLogger(__name__)
 
+app = typer.Typer(
+    name="csv",
+    add_completion=False,
+    context_settings={"help_option_names": ["-h", "--help"]},
+)
 
-def convert_to_csv(input_file, output_file):
+
+def convert_to_csv(input_file: str, output_file: str) -> None:
     in_path = Path(input_file)
     if not in_path.is_file():
         raise FileNotFoundError(f"Input file not found: {in_path}")
@@ -22,26 +29,42 @@ def convert_to_csv(input_file, output_file):
     _log.info("Data converted and saved as '%s'.", output_file)
 
 
-def main() -> None:
-    logging.basicConfig(level=logging.INFO, format="%(message)s")
-    parser = argparse.ArgumentParser(
-        description="Reads a TXT file, reorganizes the data, and converts it into a CSV file."
-    )
-    parser.add_argument(
+@app.callback(invoke_without_command=True)
+def main(
+    ctx: typer.Context,  # Контекст для нативного хелпа
+    input_file: Optional[str] = typer.Option(
+        None,
         "-i",
         "--input",
-        required=True,
         help="Path to the input TXT file. This file should contain sample names in columns and microbial taxa in rows.",
-    )
-    parser.add_argument(
+    ),
+    output_file: Optional[str] = typer.Option(
+        None,
         "-o",
         "--output",
-        required=True,
         help="Path to the output CSV file. The script will restructure the data and save it here.",
-    )
-    args = parser.parse_args()
-    convert_to_csv(args.input, args.output)
+    ),
+) -> None:
+    """Reads a TXT file, reorganizes the data, and converts it into a CSV file."""
+    logging.basicConfig(level=logging.INFO, format="%(message)s")
+
+    if input_file is None and output_file is None:
+        print(ctx.get_help())
+        raise typer.Exit()
+
+    if not input_file or not output_file:
+        print(
+            "Error: Missing required options '-i / --input' and '-o / --output'.",
+            file=sys.stderr,
+        )
+        raise typer.Exit(code=1)
+
+    try:
+        convert_to_csv(input_file, output_file)
+    except FileNotFoundError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        raise typer.Exit(code=1)
 
 
 if __name__ == "__main__":
-    main()
+    app()
