### 2026-06-01 b597b9b
feat(stats): migrate diversity CLI from argparse to typer
