### 2026-06-01 268dbde
feat(stats): migrate relabund CLI from argparse to typer
