### 2026-06-01 19fc6a3
feat(pipeline): migrate main entrypoint to typer and expose domain filters
