+++ krakenparser/pipeline.py
@@ -1,13 +1,14 @@
 #!/usr/bin/env python
 """Full KrakenParser pipeline: kreport → MPA → combined → counts → rel_abund → diversity."""
 
-import argparse
 import logging
 import shutil
 import sys
 from pathlib import Path
+from typing import Optional
 
 import pandas as pd
+import typer
 
 from krakenparser.counts.convert2csv import convert_to_csv
 from krakenparser.counts.processing_script import process_files
@@ -18,6 +19,12 @@ from krakenparser.stats.diversity import calc_alpha_div, calc_beta_div
 from krakenparser.stats.relabund import calculate_rel_abund
 
 _log = logging.getLogger(__name__)
+
+app = typer.Typer(
+    name="run",
+    add_completion=False,
+    context_settings={"help_option_names": ["-h", "--help"]},
+)
 
 
 def _is_processable(path: Path) -> bool:
@@ -38,19 +45,22 @@ _OUTPUT_SUBDIRS = ("intermediate", "counts", "rel_abund", "diversity")
 
 
 def run_pipeline(
-    input_dir: str,
-    output_dir: str | None = None,
+    input_dir: Path,
+    output_dir: Optional[Path] = None,
     keep_human: bool = False,
     viruses_only: bool = False,
+    bacteria_only: bool = False,
+    fungi_only: bool = False,
+    archaea_only: bool = False,
     rarefaction_depth: int = 1000,
-    seed: int | None = None,
+    seed: Optional[int] = None,
     overwrite: bool = False,
 ) -> None:
-    source_dir = Path(input_dir)
+    source_dir = input_dir
     if not source_dir.is_dir():
         raise FileNotFoundError(f"Input directory not found: {source_dir}")
 
-    out_dir = Path(output_dir) if output_dir else source_dir.parent
+    out_dir = output_dir if output_dir else source_dir.parent
     out_dir.mkdir(parents=True, exist_ok=True)
 
     existing = [out_dir / d for d in _OUTPUT_SUBDIRS if (out_dir / d).exists()]
@@ -68,7 +78,6 @@ def run_pipeline(
     intermediate_dir = out_dir / "intermediate"
     intermediate_dir.mkdir(exist_ok=True)
 
-    # Part 1: kreport → MPA
     mpa_dir = intermediate_dir / "mpa"
     mpa_dir.mkdir(exist_ok=True)
     for f in sorted(source_dir.iterdir()):
@@ -78,47 +87,46 @@ def run_pipeline(
             _log.info("Skipping: %s", f.name)
             continue
         out_name = f.stem + ".MPA.TXT"
-        kreport_to_mpa(str(f), str(mpa_dir / out_name), display_header=True)
+        kreport_to_mpa(f, mpa_dir / out_name, display_header=True)
 
-    # Part 2: combine MPAs
     mpa_files = sorted(mpa_dir.glob("*.MPA.TXT"))
     if not mpa_files:
-        sys.exit("Error: no MPA files found after conversion.")
+        print("Error: no MPA files found after conversion.", file=sys.stderr)
+        raise typer.Exit(code=1)
     combined_file = intermediate_dir / "COMBINED.txt"
-    combine_mpa([str(f) for f in mpa_files], str(combined_file))
+    combine_mpa(mpa_files, combined_file)
     _log.info("MPA files combined. Output: %s", combined_file)
 
-    # Part 3: split combined MPA by rank
     split_mpa(
         str(combined_file),
         str(intermediate_dir),
         keep_human=keep_human,
         viruses_only=viruses_only,
+        bacteria_only=bacteria_only,
+        fungi_only=fungi_only,
+        archaea_only=archaea_only,
     )
[…]
