### 2026-06-01 d5279eb
docs(readme): polish structure and styling
