+++ README.md
@@ -9,7 +9,9 @@
 <img src="https://github.com/PopovIILab/KrakenParser/blob/main/imgs/KrakenParser_logo_dark.png#gh-dark-mode-only" align="left"/>
 
 ## Overview
+
 KrakenParser is a collection of scripts designed to process Kraken2 reports and convert them into CSV format. This pipeline extracts taxonomic abundance data at six levels:
+
 - **Phylum**
 - **Class**
 - **Order**
@@ -17,9 +19,190 @@ KrakenParser is a collection of scripts designed to process Kraken2 reports and 
 - **Genus**
 - **Species**
 
-You can run the entire pipeline with **a single command**, or use the scripts **individually** depending on your needs.
+## Installation
+
+```bash
+# Linux / WSL / macOS
+conda create -n krakenparser pip -y
+conda activate krakenparser
+pip install krakenparser
+```
+
+## Usage Guide
+
+### Full Pipeline
+
+```bash
+KrakenParser -i data/kreports -o results/
+```
+
+This will:
+
+1. Convert Kraken2 reports to MPA format
+2. Combine MPA files into a single file
+3. Extract taxonomic levels into separate text files
+4. Process extracted text files
+5. Convert them into CSV format
+6. Calculate relative abundance
+7. Calculate α & β-diversities
+
+> [!TIP]
+> After the pipeline finishes, the output window will remind you about calibrating
+> rarefaction depth for β-diversity and re-running relative abundance normalization
+> before visualization — with ready-to-paste example commands tailored to your output paths.
+
+#### Full help output
+
+```text
+Usage: KrakenParser [OPTIONS] COMMAND [ARGS]...                                
+                                                                                
+ KrakenParser: Convert Kraken2 Reports to CSV and analyze microbial diversity.  
+                                                                                
+ To execute the full pipeline automatically, just use the global options.       
+                                                                                
+ Alternatively, you can run specific parts of the pipeline manually in the      
+ following order:                                                               
+                                                                                
+ mpa ➔ combine ➔ split ➔ process ➔ csv ➔ relabund ➔ diversity                   
+                                                                                
+ Each step behaves as an independent tool. Type 'krakenparser <command> --help' 
+ to see options for a specific step.                                            
+                                                                                
+╭─ Options ────────────────────────────────────────────────────────────────────╮
+│ --input       -i      PATH     Directory containing Kraken2 report files.    │
+│ --output      -o      PATH     Output directory.                             │
+│ --viruses                      Extract only VIRUSES domain taxa in the       │
+│                                pipeline.                                     │
+│ --bacteria                     Extract only BACTERIA domain taxa in the      │
+│                                pipeline.                                     │
+│ --fungi                        Extract only FUNGI kingdom taxa in the        │
[…]
