### 2026-06-01 58f6bb4
test(cli): update test suites to reflect modern typer subcommand syntax
