+++ tests/test_integration.py
@@ -2,6 +2,7 @@
 
 import hashlib
 import itertools
+from pathlib import Path
 
 import pandas as pd
 import pytest
@@ -39,7 +40,7 @@ def combined_mpa_file(tmp_path):
 # ---------------------------------------------------------------------------
 
 
-def _sha256(path) -> str:
+def _sha256(path: Path) -> str:
     return hashlib.sha256(path.read_bytes()).hexdigest()
 
 
@@ -53,7 +54,7 @@ def test_kreport_to_mpa_reproducible(kreport_file, tmp_path):
 
     def run():
         out = tmp_path / f"out_{next(counter)}.MPA.TXT"
-        kreport_to_mpa(str(kreport_file), str(out), display_header=True)
+        kreport_to_mpa(kreport_file, out, display_header=True)
         return _sha256(out)
 
     assert run() == run()
@@ -61,7 +62,7 @@ def test_kreport_to_mpa_reproducible(kreport_file, tmp_path):
 
 def test_kreport_to_mpa_standard_ranks_only(kreport_file, tmp_path):
     out = tmp_path / "out.MPA.TXT"
-    kreport_to_mpa(str(kreport_file), str(out))
+    kreport_to_mpa(kreport_file, out)
     lines = out.read_text().splitlines()
     paths = [ln.split("\t")[0] for ln in lines]
 
@@ -72,7 +73,7 @@ def test_kreport_to_mpa_standard_ranks_only(kreport_file, tmp_path):
 
 def test_kreport_to_mpa_excludes_unclassified_and_root(kreport_file, tmp_path):
     out = tmp_path / "out.MPA.TXT"
-    kreport_to_mpa(str(kreport_file), str(out))
+    kreport_to_mpa(kreport_file, out)
     content = out.read_text()
     assert "unclassified" not in content
     assert "root" not in content
@@ -80,7 +81,7 @@ def test_kreport_to_mpa_excludes_unclassified_and_root(kreport_file, tmp_path):
 
 def test_kreport_to_mpa_display_header(kreport_file, tmp_path):
     out = tmp_path / "out.MPA.TXT"
-    kreport_to_mpa(str(kreport_file), str(out), display_header=True)
+    kreport_to_mpa(kreport_file, out, display_header=True)
     first_line = out.read_text().splitlines()[0]
     assert first_line.startswith("#Classification")
     assert kreport_file.name in first_line
@@ -88,12 +89,11 @@ def test_kreport_to_mpa_display_header(kreport_file, tmp_path):
 
 def test_kreport_to_mpa_paths_are_hierarchical(kreport_file, tmp_path):
     out = tmp_path / "out.MPA.TXT"
-    kreport_to_mpa(str(kreport_file), str(out))
+    kreport_to_mpa(kreport_file, out)
     lines = out.read_text().splitlines()
     for ln in lines:
         path = ln.split("\t")[0]
         segments = path.split("|")
-        # Each segment must have a known two-letter prefix
         for seg in segments:
             assert "__" in seg, f"Unexpected MPA segment format: {seg!r}"
 
@@ -108,7 +108,7 @@ def test_convert_to_csv_reproducible(counts_txt_file, tmp_path):
 
     def run():
         out = tmp_path / f"out_{next(counter)}.csv"
-        convert_to_csv(str(counts_txt_file), str(out))
+        convert_to_csv(counts_txt_file, out)
         return _sha256(out)
 
     assert run() == run()
@@ -116,7 +116,7 @@ def test_convert_to_csv_reproducible(counts_txt_file, tmp_path):
 
 def test_convert_to_csv_transposes_correctly(counts_txt_file, tmp_path):
[…]
