### 2026-06-01 36adecd
chore: remove obsolete comments from scripts
