### 2026-06-02 65fae17
docs(readme): update CLI help usage with new single-dash aliases
