### 2026-06-03 796d5ef
docs(paper): update manuscript text to reflect current features and design
