+++ paper/paper.md
@@ -13,41 +13,47 @@ affiliations:
  - name: Faculty of Bioengineering and Veterinary Medicine, Don State Technical University, Russia
    index: 1
    ror: 00x5je630
-date: 1 June 2025
+date: 3 June 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-`KrakenParser` is an open-source software tool (with a command-line interface and Python API) designed to streamline the post-analysis of metagenomic classification results produced by `Kraken2` [@wood2019kraken2] and similar taxonomic profilers such as `Bracken` [@lu2017bracken] and `Metabuli` [@kim2024metabuli]. `Kraken2` is a widely used taxonomic classifier that assigns metagenomic reads to taxa using exact k-mer matches, achieving high speed and accuracy. However, the raw output of `Kraken2` [@wood2019kraken2] (and related tools) is a text report that can be cumbersome to interpret and aggregate across multiple samples. `KrakenParser` addresses this need by converting multiple `Kraken`-format reports into structured tables (CSV files) at various taxonomic ranks (from phylum down to species), performing filtering and normalization (including relative abundance calculations), and providing APIs to produce publication-ready plots. The tool automates the multi-step process of combining and cleaning `Kraken` results, allowing researchers to quickly obtain human-readable summaries of community composition. `KrakenParser`’s focus is on efficiency, ease-of-use, and integration: it can run an entire conversion pipeline with a single command and also be imported as a Python library for custom workflows. In summary, `KrakenParser` significantly reduces the manual effort required to post-process metagenomic classification data, enabling scientists to go from raw classifier output to analysis-ready tables and figures in one step.
+`KrakenParser` is an open-source Python library and command-line interface (CLI) designed to automate andคู่ standardize the post-processing of metagenomic taxonomic classification reports generated by `Kraken2` [@wood2019kraken2], `Bracken` [@lu2017bracken], and `Metabuli` [@kim2024metabuli]. Although these high-throughput profilers provide rapid taxonomic assignments, their primary outputs comprise independent, hierarchical text files that are challenging to aggregate, filter, and analyze across multi-sample cohorts. `KrakenParser` addresses this bottleneck by parsing and merging separate report files into unified, rank-specific (from phylum down to species) count matrices formatted as tidy CSV tables. Additionally, the tool provides advanced taxonomic filtering, normalizes raw counts into relative abundances, computes ecological alpha and beta diversity metrics, and offers an object-oriented API for publication-ready visualizations. `KrakenParser` can be deployed as an automated end-to-end pipeline via a single CLI command or integrated as a modular library within custom bioinformatics workflows.
 
 # Statement of need
 
[…]
