### 2026-06-03 96ab075
docs(paper): add statistical and ecological references to bibliography
