+++ paper/paper.bib
@@ -148,3 +148,67 @@
   doi       = {10.21105/joss.03021},
   url       = {https://doi.org/10.21105/joss.03021}
 }
+
+@article{shannon1948mathematical,
+  title={A mathematical theory of communication},
+  author={Shannon, Claude Elwood},
+  journal={The Bell system technical journal},
+  volume={27},
+  number={3},
+  pages={379--423},
+  year={1948},
+  publisher={Nokia Bell Labs}
+}
+
+@article{pielou1966measurement,
+  title={The measurement of diversity in different types of biological collections},
+  author={Pielou, Evelyn C},
+  journal={Journal of theoretical biology},
+  volume={13},
+  pages={131--144},
+  year={1966},
+  publisher={Elsevier}
+}
+
+@article{chao2002estimating,
+  title={Estimating the number of species in a stochastic abundance model},
+  author={Chao, Anne and Bunge, John},
+  journal={Biometrics},
+  volume={58},
+  number={3},
+  pages={531--539},
+  year={2002},
+  publisher={Wiley Online Library}
+}
+
+@article{bray10jt,
+  title={JT Curtis An ordination of the upland forest communities of southern Wisconsin., 1957, 27},
+  author={Bray, JR},
+  journal={DOI: https://doi. org/10.2307/1942268},
+  pages={325--349}
+}
+
+@article{jaccard1901etude,
+  title={{\'E}tude comparative de la distribution florale dans une portion des Alpes et des Jura},
+  author={Jaccard, Paul},
+  journal={Bull Soc Vaudoise Sci Nat},
+  volume={37},
+  pages={547--579},
+  year={1901}
+}
+
+
+@article{ijms26135941,
+  AUTHOR = {Popov, Ilia V. and Manakhov, Andrey D. and Gorobets, Vladislav E. and Diakova, Kristina B. and Lukbanova, Ekaterina A. and Malinovkin, Aleksey V. and Venema, Koen and Ermakov, Alexey M. and Popov, Igor V.},
+  TITLE = {Metagenomic Investigation of Intestinal Microbiota of Insectivorous Synanthropic Bats: Densoviruses, Antibiotic Resistance Genes, and Functional Profiling of Gut Microbial Communities},
+  JOURNAL = {International Journal of Molecular Sciences},
+  VOLUME = {26},
+  YEAR = {2025},
+  NUMBER = {13},
+  ARTICLE-NUMBER = {5941},
+  URL = {https://www.mdpi.com/1422-0067/26/13/5941},
+  PubMedID = {40649719},
+  ISSN = {1422-0067},
[…]
