### 2026-06-03 6d747c9
docs(paper): fix CLI execution description and streamline text
