### 2026-06-03 2b619a7
docs(paper): polish paper metadata and sections
