### 2026-06-04 58c25c0
docs(cli): add type annotations, docstrings, and comments to krakenparser entrypoint
