### 2026-06-04 fa0a081
refactor(pipeline): harden pipeline orchestration with static type hints and docstrings
