+++ krakenparser/pipeline.py
@@ -1,5 +1,10 @@
-#!/usr/bin/env python
-"""Full KrakenParser pipeline: kreport → MPA → combined → counts → rel_abund → diversity."""
+#!/usr/bin/env python3
+"""Core orchestration engine for the KrakenParser execution pipeline.
+
+This module consolidates independent taxonomic processing steps into a seamless,
+end-to-end automated pipeline. It handles file validation, directory structure
+sanitization, data transformations, statistical scaling, and diversity indexing.
+"""
 
 import logging
 import shutil
@@ -18,30 +23,52 @@ from krakenparser.mpa.transform2mpa import kreport_to_mpa
 from krakenparser.stats.diversity import calc_alpha_div, calc_beta_div
 from krakenparser.stats.relabund import calculate_rel_abund
 
-_log = logging.getLogger(__name__)
+# Initialize module-level isolated logger
+_log: logging.Logger = logging.getLogger(__name__)
 
-app = typer.Typer(
+# Primary pipeline subcommand routing sub-app
+app: typer.Typer = typer.Typer(
     name="run",
     add_completion=False,
     context_settings={"help_option_names": ["-h", "--help"]},
 )
 
+# Structural database directory map enforced by the pipeline architecture
+_OUTPUT_SUBDIRS: tuple[str, ...] = ("intermediate", "counts", "rel_abund", "diversity")
 
-def _is_processable(path: Path) -> bool:
-    """Return False for hidden files, files with null bytes, or non-UTF-8 files."""
-    if path.name.startswith("."):
+
+def _is_processable(filepath: Path) -> bool:
+    """Validate file integrity prior to feeding it to text-parsing utilities.
+
+    Ensures the target file is not a hidden system artifact, does not contain
+    binary null-byte contamination, and safely decodes via UTF-8 standard.
+
+    Args:
+        path: A Path object referencing the target text file.
+
+    Returns:
+        bool: True if the file matches structural sanity baselines, False otherwise.
+    """
+    if filepath.name.startswith("."):
         return False
+
     try:
-        chunk = path.read_bytes()[:1024]
-        if b"\x00" in chunk:
+        if b"\x00" in filepath.read_bytes():
             return False
-        chunk.decode("utf-8")
-        return True
-    except (UnicodeDecodeError, OSError):
+    except Exception:
         return False
 
-
-_OUTPUT_SUBDIRS = ("intermediate", "counts", "rel_abund", "diversity")
+    # 3. СТРОГАЯ проверка на UTF-8 (вот здесь косяк)
+    try:
+        # Добавляем errors="strict", чтобы не глотать BOM и левые кодировки
+        with open(filepath, "r", encoding="utf-8", errors="strict") as f:
+            # Читаем небольшой кусок файла для проверки
+            f.read(1024)
+        return True
+    except UnicodeDecodeError:
+        return False
+    except Exception:
+        return False
 
 
 def run_pipeline(
@@ -56,16 +83,43 @@ def run_pipeline(
     seed: Optional[int] = None,
     overwrite: bool = False,
 ) -> None:
-    source_dir = input_dir
+    """Execute the sequential programmatic workflow blocks of KrakenParser.
+
+    Validates resources, purges legacy directories under explicit overwrite rules,
+    converts raw reports to structured matrices, strips structural taxonomic strings,
+    scales measurements, and exports ecological diversity indices.
+
+    Args:
+        input_dir: Path pointing to the source directory containing raw reports.
+        output_dir: Custom export destination path. Defaults to input_dir parent if None.
+        keep_human: If True, biological filtering blocks skipping human contamination are disabled.
[…]
