+++ krakenparser/__init__.py
@@ -2,7 +2,7 @@ from .kpplot.clustermap import clustermap
 from .kpplot.stackedbar import stacked_barplot
 from .kpplot.streamgraph import streamgraph
 
-__all__ = [
+__all__: list[str] = [
     "stacked_barplot",
     "streamgraph",
     "clustermap",
