### 2026-06-04 83f155a
refactor(counts): modernize matrix transposition utility with Path typing and docstrings
