### 2026-06-04 7edb694
refactor(counts): add type annotations and docstrings to processing_script.py
