### 2026-06-04 3824525
refactor(counts): add type annotations and docstrings to split_mpa.py
