+++ krakenparser/counts/split_mpa.py
@@ -1,7 +1,10 @@
-#!/usr/bin/env python
-"""Split a combined MPA table into per-rank TXT files.
+#!/usr/bin/env python3
+"""Decomposition utility to partition master MPA matrices into rank-specific text tables.
 
-Replaces decombine.sh and decombine_viruses.sh.
+This module splits combined multi-sample MetaPhlAn files into separate tables grouped
+by taxonomic rank (species, genus, family, etc.). It supports on-the-fly filtering
+for specific biological domains (e.g., Viruses, Bacteria) and filters out host
+contamination profiles using predefined taxonomic blacklists.
 """
 
 import logging
@@ -14,15 +17,18 @@ import typer
 
 from krakenparser.utils import ensure_output_dir
 
-_log = logging.getLogger(__name__)
+# Initialize module-level isolated logger
+_log: logging.Logger = logging.getLogger(__name__)
 
-app = typer.Typer(
+# Dedicated Typer routing application instantiation
+app: typer.Typer = typer.Typer(
     name="split",
     add_completion=False,
     context_settings={"help_option_names": ["-h", "--help"]},
 )
 
-_RANKS = [
+# Immutable configuration schema mapping ranks, targets, and descendents to drop
+_RANKS: list[tuple[str, str, list[str]]] = [
     ("species", "s__", []),
     ("genus", "g__", ["s__"]),
     ("family", "f__", ["s__", "g__"]),
@@ -31,7 +37,8 @@ _RANKS = [
     ("phylum", "p__", ["s__", "g__", "f__", "o__", "c__"]),
 ]
 
-_HUMAN_MARKERS = frozenset(
+# Host/human filtering target taxonomy markers
+_HUMAN_MARKERS: frozenset[str] = frozenset(
     [
         "s__Homo_sapiens",
         "g__Homo",
@@ -42,48 +49,88 @@ _HUMAN_MARKERS = frozenset(
     ]
 )
 
-_ACCESSION_RE = re.compile(r"(SRS|SRR|SRX|ERS|ERR|ERX|DRS|DRR|DRX)\d*-")
+# Regular expression matching SRA/ENA run technical accession sub-strings
+_ACCESSION_RE: re.Pattern[str] = re.compile(
+    r"(SRS|SRR|SRX|ERS|ERR|ERX|DRS|DRR|DRX)\d*-"
+)
 
 
 def _strip_path_prefix(line: str) -> str:
-    tab = line.find("\t")
+    """Isolate the terminal taxonomic clade and purge short-read archive prefixes.
+
+    Extracts the right-most node classification component from pipe-separated
+    lineage string paths and trims technical sequencing metadata tags.
+
+    Args:
+        line: Raw tab-separated line from an MPA summary matrix.
+
+    Returns:
+        str: Cleansed and isolated clade metric description row.
+    """
+    tab: int = line.find("\t")
     if tab == -1:
         return line
     path, rest = line[:tab], line[tab:]
-    pipe = path.rfind("|")
-    segment = path[pipe + 1 :] if pipe != -1 else path
+    pipe: int = path.rfind("|")
+    segment: str = path[pipe + 1 :] if pipe != -1 else path
     return _ACCESSION_RE.sub("", segment + rest)
 
 
 def _human_in_line(line: str) -> bool:
-    tab = line.find("\t")
-    path = line[:tab] if tab != -1 else line
-    segments = set(path.split("|"))
+    """Verify if the taxonomic lineage contains human contamination markers.
+
+    Args:
+        line: Raw text line containing structural pipe-separated classifications.
+
+    Returns:
+        bool: True if the lineage intersects with monitored human host constraints.
+    """
+    tab: int = line.find("\t")
[…]
