### 2026-06-04 f190b3e
refactor(kpplot): expand sub-package initialization namespace and enforce explicit export typing
