+++ krakenparser/kpplot/__init__.py
@@ -1,3 +1,14 @@
 from .base import KpPlotBase
+from .clustermap import KpClustermap, clustermap
+from .stackedbar import KpStackedBarplot, stacked_barplot
+from .streamgraph import KpStreamgraph, streamgraph
 
-__all__ = ["KpPlotBase"]
+__all__: list[str] = [
+    "KpPlotBase",
+    "KpClustermap",
+    "clustermap",
+    "KpStackedBarplot",
+    "stacked_barplot",
+    "KpStreamgraph",
+    "streamgraph",
+]
