### 2026-06-04 8989fb2
refactor(kpplot): implement static type hinting and docstrings in plotting base
