+++ krakenparser/kpplot/base.py
@@ -1,37 +1,61 @@
-from typing import Optional
+#!/usr/bin/env python3
+"""Base classes and data aggregation utilities for metagenomic visualization.
+
+This module provides structural baselines for managing Matplotlib canvas states
+and executing dataframe transformations, such as merging abundance matrices
+with cohort metadata schemas and performing group-level re-normalization.
+"""
+
+from pathlib import Path
+from typing import Optional, Union
 
 import matplotlib.pyplot as plt
 import pandas as pd
 
 
 class KpPlotBase:
-    def __init__(self, fig: plt.Figure, ax: plt.Axes):
-        self.fig = fig
-        self.ax = ax
+    """Base orchestration class managing Matplotlib figure and axis contexts."""
+
+    def __init__(self, fig: plt.Figure, ax: plt.Axes) -> None:
+        """Initialize the plot base wrapper with canvas references.
+
+        Args:
+            fig: The Matplotlib Figure instance acting as the global canvas.
+            ax: The Matplotlib Axes instance mapping sub-plot coordinate space.
+        """
+        self.fig: plt.Figure = fig
+        self.ax: plt.Axes = ax
 
     def plotfig(self) -> plt.Figure:
-        # self.fig.tight_layout()
+        """Render the underlying canvas interactively.
+
+        Returns:
+            plt.Figure: The updated Matplotlib Figure object.
+        """
         plt.show()
         return self.fig
 
     def savefig(
         self,
-        path: str,
+        path: Union[Path, str],
         dpi: int = 300,
         transparent: bool = False,
         bbox_inches: Optional[str] = "tight",
-    ):
-        """
-        Save the figure to a file.
+    ) -> None:
+        """Commit the current canvas state atomically to a physical image file.
 
-        Parameters:
-        - path: Path to save the figure (e.g. "plot.png", "plot.svg").
-        - dpi: Dots per inch (resolution) of the output image. Default is 300.
-        - transparent: Whether to make saved figure transparent.
-        - bbox_inches: Bounding box option passed to matplotlib. Default is "tight".
+        Args:
+            path: Target location or string path where the layout image is exported.
+            dpi: Dots per inch managing resolution limits. Defaults to 300.
+            transparent: Toggles transparency anchors for background regions.
+            bbox_inches: Boundary box padding constraints. Defaults to "tight".
         """
+        target_path: Path = Path(path)
         self.fig.savefig(
-            path, dpi=dpi, transparent=transparent, bbox_inches=bbox_inches
+            target_path,
+            dpi=dpi,
+            transparent=transparent,
+            bbox_inches=bbox_inches,
         )
 
 
@@ -40,18 +64,46 @@ def aggregate_by_metadata(
     metadata: pd.DataFrame,
     metadata_group: str,
 ) -> pd.DataFrame:
-    """Merge df with metadata and re-normalise rel_abund_perc per group."""
+    """Consolidate abundance samples by cohorts and re-normalize relative profiles.
+
+    Maps discrete sample rows to experimental variables, calculates group-specific
+    abundance means per taxon, and scales profiles to ensure the sum equals 100%.
+
+    Args:
+        df: Input DataFrame containing 'Sample_id', 'taxon', and 'rel_abund_perc'.
+        metadata: Metadata worksheet mapping 'Sample_id' to cohort traits.
+        metadata_group: Specific targeted feature column label inside metadata.
+
+    Returns:
+        pd.DataFrame: A transposed tidy dataframe ready for cohort-wide plotting.
+
+    Raises:
+        ValueError: Triggered if 'Sample_id' or 'metadata_group' columns are missing.
+    """
[…]
