### 2026-06-04 c8aea1f
refactor(kpplot): add strict literal constraints, type annotations, and docstrings to stackedbar.py
