+++ krakenparser/kpplot/stackedbar.py
@@ -1,4 +1,12 @@
-from typing import List, Optional, Tuple, Union
+#!/usr/bin/env python3
+"""Stacked barplot visualization module for compositional metagenomic profiles.
+
+This module builds normalized, multi-component stacked column charts tracking
+relative taxonomic abundance variations across independent sample coordinates
+or aggregated experimental cohorts.
+"""
+
+from typing import Literal, Optional, Sequence, Tuple, Union
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -9,132 +17,174 @@ from .base import KpPlotBase, aggregate_by_metadata
 
 
 class KpStackedBarplot(KpPlotBase):
-    pass
+    """Orchestration context wrapper encapsulating Matplotlib stacked barplot layouts."""
+
+    def __init__(self, fig: plt.Figure, ax: plt.Axes) -> None:
+        """Initialize the stacked barplot canvas with layout metrics.
+
+        Args:
+            fig: The Matplotlib Figure container hosting the drawing canvas.
+            ax: The core underlying Axes coordinate grid mapper.
+        """
+        super().__init__(fig, ax)
 
 
 def stacked_barplot(
-    df,
+    df: pd.DataFrame,
     metadata: Optional[pd.DataFrame] = None,
     metadata_group: Optional[str] = None,
-    sample_order: Optional[List[str]] = None,
-    figsize: Tuple[int, int] = (14, 7),
-    cmap: Optional[Union[str, List[str]]] = "tab20",
+    sample_order: Optional[Sequence[str]] = None,
+    figsize: Tuple[float, float] = (14.0, 7.0),
+    cmap: Union[str, Sequence[str]] = "tab20",
     bar_width: float = 0.6,
     edgecolor: str = "black",
     edge_linewidth: float = 0.3,
     title: Optional[str] = None,
     title_fontsize: float = 16.0,
     title_color: str = "black",
-    title_weight: str = "normal",
-    title_style: str = "normal",
+    title_weight: Literal["normal", "bold", "heavy", "light"] = "normal",
+    title_style: Literal["normal", "italic", "oblique"] = "normal",
     xlabel: str = "Samples",
     xlabel_fontsize: float = 12.0,
     xlabel_color: str = "black",
-    xlabel_weight: str = "normal",
-    xlabel_style: str = "normal",
+    xlabel_weight: Literal["normal", "bold", "heavy", "light"] = "normal",
+    xlabel_style: Literal["normal", "italic", "oblique"] = "normal",
     ylabel: str = "Relative Abundance (%)",
     ylabel_fontsize: float = 12.0,
     ylabel_color: str = "black",
-    ylabel_weight: str = "normal",
-    ylabel_style: str = "normal",
+    ylabel_weight: Literal["normal", "bold", "heavy", "light"] = "normal",
+    ylabel_style: Literal["normal", "italic", "oblique"] = "normal",
     xticks_rotation: float = 0.0,
-    xticks_ha: str = "center",
+    xticks_ha: Literal["left", "right", "center"] = "center",
     xticks_fontsize: float = 12.0,
     xticks_color: str = "black",
-    xticks_weight: str = "normal",
-    xticks_style: str = "normal",
-    background_color="white",
+    xticks_weight: Literal["normal", "bold", "heavy", "light"] = "normal",
+    xticks_style: Literal["normal", "italic", "oblique"] = "normal",
+    background_color: Optional[str] = "white",
     grid: bool = True,
     grid_linestyle: str = "--",
     grid_alpha: float = 0.7,
     legend_title: str = "Taxon",
     legend_fontsize: float = 9.0,
-    legend_fontstyle: str = "italic",
[…]
