### 2026-06-04 b5dcee4
refactor(kpplot): add strict literal constraints, type annotations, and docstrings to streamgraph.py
