### 2026-06-04 fe6464d
refactor(mpa): implement explicit type hints, and docstrings in mpa_table.py
