+++ krakenparser/mpa/mpa_table.py
@@ -1,5 +1,11 @@
-#!/usr/bin/env python
-"""Combine multiple MPA-format files into a single merged table."""
+#!/usr/bin/env python3
+"""Aggregation engine for merging multiple MetaPhlAn (MPA) files into a unified matrix.
+
+This module parses multi-sample taxonomic report files formatted in the MetaPhlAn
+style, extracts their respective abundance sequences, tracks phylogenetic tree
+parent-child relationships, and performs a stack-based traversal to output a
+structurally ordered, tab-delimited master count matrix.
+"""
 
 import logging
 import sys
@@ -10,9 +16,11 @@ import typer
 
 from krakenparser.utils import ensure_output_dir
 
-_log = logging.getLogger(__name__)
+# Initialize module-level isolated logger
+_log: logging.Logger = logging.getLogger(__name__)
 
-app = typer.Typer(
+# Dedicated Typer routing application instantiation
+app: typer.Typer = typer.Typer(
     name="combine",
     add_completion=False,
     context_settings={"help_option_names": ["-h", "--help"]},
@@ -20,31 +28,49 @@ app = typer.Typer(
 
 
 def combine_mpa(in_files: list[Path], o_file: Path) -> None:
-    out_path = ensure_output_dir(str(o_file), is_file=True)
+    """Merge separate MPA taxonomic distribution tables into a single master matrix.
 
+    Parses header metadata strings to resolve human-readable sample IDs, caches
+    hierarchical dependencies to maintain strict lineage orientation, and combines
+    individual abundance values into an integrated layout table.
+
+    Args:
+        in_files: List of validated Path objects directing to sample MPA files.
+        o_file: Target Path where the combined tab-delimited table will be written.
+
+    Raises:
+        FileNotFoundError: Triggered if any referenced input file is absent.
+    """
+    out_path: Path = ensure_output_dir(o_file, is_file=True)
+
+    # Architectural storage definitions for parsing alignment graphs
     samples: dict[int, str] = {}
     values: dict[str, dict[int, str]] = {}
     parent2child: dict[str, list[str]] = {}
     toparse: list[str] = []
-    sample_count = 0
+    sample_count: int = 0
 
     _log.info("Number of files to parse: %d", len(in_files))
 
+    # Atomic verification step protecting IO transactions
     for in_path in in_files:
         if not in_path.is_file():
             raise FileNotFoundError(f"Input file not found: {in_path}")
 
+    # Step 1: Scan individual reports and populate hierarchical graphs
     for in_path in in_files:
         sample_count += 1
-        sample_name = f"Sample #{sample_count}"
+        sample_name: str = f"Sample #{sample_count}"
 
-        with open(in_path) as fh:
+        with open(in_path, encoding="utf-8") as fh:
             for line in fh:
                 line = line.rstrip("\n")
                 if not line:
                     continue
+
+                # Isolate and extract sample identifier mapping embedded within headers
                 if line.startswith("#"):
-                    cols = line.split("\t")
+                    cols: list[str] = line.split("\t")
                     if len(cols) >= 2:
                         sample_name = cols[-1]
                     continue
@@ -54,10 +80,11 @@ def combine_mpa(in_files: list[Path], o_file: Path) -> None:
                     continue
                 classification, val = cols[0], cols[1]
 
-                split_vals = classification.split("|")
-                curr_parent = ""
+                # Resolve lineage parent node identities to guarantee structural order
+                split_vals: list[str] = classification.split("|")
+                curr_parent: str = ""
                 for i in range(len(split_vals)):
-                    test_val = "|".join(split_vals[:i])
+                    test_val: str = "|".join(split_vals[:i])
[…]
