### 2026-06-04 0560267
refactor(mpa): add strict type annotations, context managers, and docstrings to transform2mpa.py
