+++ krakenparser/mpa/transform2mpa.py
@@ -1,51 +1,64 @@
-#!/usr/bin/env python
-"""Convert a Kraken2 report to MetaPhlAn (MPA) format."""
+#!/usr/bin/env python3
+"""Taxonomic format converter translating Kraken2 reports to MetaPhlAn (MPA) layout.
+
+This module parses standard space-indented hierarchical Kraken2 and KrakenUniq output
+reports, tracks taxonomic depth changes through parent lineage state machines,
+and converts records into pipe-separated '|' multi-level lineage tracks.
+"""
 
 import logging
-import os
 import sys
 from pathlib import Path
-from typing import Optional
+from typing import Any, Optional
 
 import typer
 
 from krakenparser.utils import ensure_output_dir
 
-_log = logging.getLogger(__name__)
+# Initialize module-level isolated logger
+_log: logging.Logger = logging.getLogger(__name__)
 
-app = typer.Typer(
+# Dedicated Typer routing application instantiation
+app: typer.Typer = typer.Typer(
     name="mpa",
     add_completion=False,
     context_settings={"help_option_names": ["-h", "--help"]},
 )
 
-_MAIN_LVLS = {"R", "K", "D", "P", "C", "O", "F", "G", "S"}
+# Standard strict taxonomic reference limits
+_MAIN_LVLS: set[str] = {"R", "K", "D", "P", "C", "O", "F", "G", "S"}
 
 
 def _parse_line(line: str, remove_spaces: bool = False) -> list:
-    """Parse one Kraken2 report line.
+    """Parse a single Kraken2 or KrakenUniq report row and extract vital stats.
 
-    Returns [name, level_num, level_type, all_reads, percents]
-    or empty list on malformed input.
+    Handles standard kraken formats alongside KrakenUniq outputs by mapping text
+    labels to fixed-width single-character taxonomic rank designators.
+
+    Args:
+        line: A raw tab-delimited line from the report file.
+        remove_spaces: If True, internal spaces within organism nomenclature
+            strings are mapped securely to structural underscores.
+
+    Returns:
+        list: A data list containing [cleaned_name, level_num, level_type, total_reads, relative_percentage]
+            or an empty list [] if the line format violates parser syntactic assumptions.
     """
-    parts = line.rstrip("\n").split("\t")
+    parts: list[str] = line.rstrip("\n").split("\t")
     if len(parts) < 4:
         return []
+
     try:
-        int(parts[1])
+        percents: float = float(parts[0])
+        all_reads: int = int(parts[1])
     except ValueError:
         return []
 
-    try:
-        percents = float(parts[0])
-    except ValueError:
-        return []
-    all_reads = int(parts[1])
-
+    # Detect and handle alternative KrakenUniq columns format if applicable
     try:
         int(parts[-3])
-        level_type = parts[-2].strip()
-        map_kuniq = {
+        level_type: str = parts[-2].strip()
+        map_kuniq: dict[str, str] = {
             "species": "S",
             "genus": "G",
             "family": "F",
@@ -55,25 +68,20 @@ def _parse_line(line: str, remove_spaces: bool = False) -> list:
             "superkingdom": "D",
             "kingdom": "K",
         }
-        if level_type not in map_kuniq:
-            level_type = "-"
-        else:
-            level_type = map_kuniq[level_type]
+        level_type = map_kuniq.get(level_type, "-")
     except ValueError:
         level_type = parts[-3].strip()
 
-    name = parts[-1]
-    spaces = 0
-    for ch in name:
-        if ch == " ":
-            spaces += 1
-        else:
-            break
-    name = name.strip()
+    raw_name: str = parts[-1]
+
[…]
