### 2026-06-04 1116b04
refactor(stats): implement vectorization, explicit type hints, and docstrings in relabund.py
