### 2026-06-04 a639f50
refactor(tests): parameterize CLI error paths and clean up entrypoint testing in test_cli.py
