+++ tests/test_cli.py
@@ -1,10 +1,19 @@
-"""Smoke tests for CLI entry-points via Typer CliRunner."""
+"""Smoke tests for CLI entry-points via Typer CliRunner.
+
+Each command is exercised for three standard error paths:
+  1. No arguments → usage help (exit 0).
+  2. Missing one required option → validation error (exit 1).
+  3. Non-existent input file → runtime error (exit 1).
+
+Happy paths for commands that produce file output are covered separately in
+test_integration.py and test_full_pipeline.py.
+"""
 
 import shutil
 
 import pandas as pd
 import pytest
-from typer.testing import CliRunner
+from conftest import SAMPLE_MPA_A, SAMPLE_MPA_B
 
 from krakenparser.counts.convert2csv import app as convert2csv_app
 from krakenparser.counts.processing_script import app as processing_app
@@ -15,9 +24,6 @@ from krakenparser.pipeline import app as pipeline_app
 from krakenparser.stats.diversity import app as diversity_app
 from krakenparser.stats.relabund import app as relabund_app
 
-_MPA_A = "#Classification\tsample1\nd__Bacteria|s__Pseudomonas_aeruginosa\t300\n"
-_MPA_B = "#Classification\tsample2\nd__Bacteria|s__Pseudomonas_aeruginosa\t100\n"
-
 _COMBINED_MPA = (
     "#Classification\tsample1\tsample2\n"
     "d__Bacteria|p__Pseudomonadota|g__Pseudomonas|s__Pseudomonas_aeruginosa\t300\t100\n"
@@ -25,118 +31,102 @@ _COMBINED_MPA = (
 )
 
 
-@pytest.fixture
-def runner():
-    return CliRunner()
-
-
 # ---------------------------------------------------------------------------
-# convert2csv
+# Parametrized standard error-path suite
 # ---------------------------------------------------------------------------
 
-
-def test_convert2csv_no_args_shows_help(runner):
-    result = runner.invoke(convert2csv_app, [])
-    assert result.exit_code == 0
-    assert "Usage" in result.output
-
-
-def test_convert2csv_missing_one_option(runner, tmp_path):
-    result = runner.invoke(convert2csv_app, ["-i", str(tmp_path / "x.txt")])
-    assert result.exit_code == 1
-    assert "Missing required options" in result.output
-
-
-def test_convert2csv_file_not_found(runner, tmp_path):
-    result = runner.invoke(
+# Each tuple: (app, no-arg exit code, partial-args list, file-not-found args factory)
+# The factory is a callable(tmp_path) → list[str].
+_CLI_SPECS = [
+    (
         convert2csv_app,
-        ["-i", str(tmp_path / "ghost.txt"), "-o", str(tmp_path / "out.csv")],
-    )
-    assert result.exit_code == 1
-    assert "Error" in result.output
-
-
-# ---------------------------------------------------------------------------
-# processing_script
-# ---------------------------------------------------------------------------
-
-
-def test_processing_no_args_shows_help(runner):
-    result = runner.invoke(processing_app, [])
-    assert result.exit_code == 0
-    assert "Usage" in result.output
-
-
-def test_processing_missing_one_option(runner, tmp_path):
-    result = runner.invoke(processing_app, ["-i", str(tmp_path / "x.txt")])
-    assert result.exit_code == 1
-    assert "Missing required options" in result.output
-
-
-def test_processing_file_not_found(runner, tmp_path):
-    result = runner.invoke(
+        ["-i", "{{ghost}}"],
+        ["-i", "{{ghost}}", "-o", "{{out}}"],
+    ),
+    (
         processing_app,
[…]
