### 2026-06-04 583d6a8
refactor(tests): group logic flows and add descriptive assertions to test_units.py
