+++ tests/test_units.py
@@ -1,4 +1,9 @@
-"""Pure-function unit tests — no I/O, fully deterministic."""
+"""Pure-function unit tests — no I/O, no fixtures, fully deterministic.
+
+All functions under test are stateless mathematical transforms or text
+processors.  Tests are grouped by function and ordered from simplest
+(zero/one element) to more complex (multi-element, edge cases).
+"""
 
 import math
 from pathlib import Path
@@ -11,14 +16,15 @@ from krakenparser.mpa.transform2mpa import _parse_line
 from krakenparser.stats.diversity import chao1_index, pielou_evenness, shannon_index
 from krakenparser.utils import ensure_output_dir
 
-# ---------------------------------------------------------------------------
+# ===========================================================================
 # _parse_line
-# ---------------------------------------------------------------------------
+# ===========================================================================
 
 
 def test_parse_line_standard_rank():
     line = "50.00\t500000\t100000\tP\t1224\t    Pseudomonadota\n"
     name, depth, rank, cum_reads, pct = _parse_line(line)
+
     assert name == "Pseudomonadota"
     assert depth == 2  # 4 leading spaces // 2
     assert rank == "P"
@@ -26,9 +32,10 @@ def test_parse_line_standard_rank():
     assert pct == 50.0
 
 
-def test_parse_line_root_no_indent():
+def test_parse_line_root_has_zero_depth():
     line = "99.98\t999980\t0\tR\t1\troot\n"
     name, depth, rank, cum_reads, pct = _parse_line(line)
+
     assert name == "root"
     assert depth == 0
     assert rank == "R"
@@ -37,47 +44,48 @@ def test_parse_line_root_no_indent():
 def test_parse_line_intermediate_rank():
     line = "5.00\t50000\t0\tS1\t12345\t          Some subspecies\n"
     name, depth, rank, _, _ = _parse_line(line)
+
     assert name == "Some subspecies"
     assert rank == "S1"
     assert depth == 5  # 10 spaces // 2
 
 
-def test_parse_line_too_few_columns():
+def test_parse_line_too_few_columns_returns_empty():
     assert _parse_line("50.00\t500000\n") == []
 
 
-def test_parse_line_non_numeric_pct():
+def test_parse_line_non_numeric_pct_returns_empty():
     assert _parse_line("not_a_float\t500000\t0\tP\t1224\tBacteria\n") == []
 
 
-def test_parse_line_non_numeric_reads():
+def test_parse_line_non_numeric_reads_returns_empty():
     assert _parse_line("50.00\tnot_int\t0\tP\t1224\tBacteria\n") == []
 
 
-# ---------------------------------------------------------------------------
+# ===========================================================================
 # shannon_index
-# ---------------------------------------------------------------------------
-
-
-def test_shannon_uniform_four_species():
-    assert abs(shannon_index([1, 1, 1, 1]) - math.log(4)) < 1e-10
+# ===========================================================================
 
 
 def test_shannon_single_species_is_zero():
     assert shannon_index([100]) == pytest.approx(0.0)
 
 
-def test_shannon_ignores_zero_counts():
-    assert abs(shannon_index([1, 1, 1, 1, 0, 0]) - math.log(4)) < 1e-10
-
-
 def test_shannon_two_equal_species():
     assert abs(shannon_index([50, 50]) - math.log(2)) < 1e-10
 
 
-# ---------------------------------------------------------------------------
+def test_shannon_uniform_four_species():
+    assert abs(shannon_index([1, 1, 1, 1]) - math.log(4)) < 1e-10
+
+
+def test_shannon_ignores_zero_counts():
[…]
