### 2026-06-04 e8f0c9c
docs(paper): add disclosures section to paper.md
