### 2026-06-04 bcbe409
docs(paper): fix layout spacing in paper.md
