### 2026-06-18 4901fec
style(toml): format pyproject.toml with Even Better TOML
