+++ pyproject.toml
@@ -6,9 +6,9 @@ build-backend = "setuptools.build_meta"
 name = "krakenparser"
 version = "1.1.2"
 description = "A collection of scripts designed to process Kraken2 reports and convert them into CSV format."
-readme = {file = "README.md", content-type = "text/markdown"}
-license = {file = "LICENSE"}
-authors = [{name = "Ilia Popov", email = "iljapopov17@gmail.com"}]
+readme = { file = "README.md", content-type = "text/markdown" }
+license = { file = "LICENSE" }
+authors = [{ name = "Ilia Popov", email = "iljapopov17@gmail.com" }]
 requires-python = ">=3.10,<=3.16"
 dependencies = [
     "pandas>=2.0.0,<=2.3.3",
@@ -20,10 +20,7 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-dev = [
-    "pytest>=7.0.0",
-    "pytest-cov>=4.0.0",
-]
+dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]
 
 [project.urls]
 Homepage = "https://github.com/PopovIILab/KrakenParser"
@@ -52,7 +49,4 @@ omit = [
 ]
 
 [tool.coverage.report]
-exclude_lines = [
-    "pragma: no cover",
-    "if __name__ == .__main__.:",
-]
+exclude_lines = ["pragma: no cover", "if __name__ == .__main__.:"]
