### 2026-08-28 #9694 @ilypopv
Hey!
Thank you for the detailed and constructive review. I've addressed all nine points in the latest [Pull Request](https://github.com/PopovIILab/KrakenParser/pull/26)!

### 1. Supported input formats
Bracken re-estimates abundances and writes them back out in the same Kraken2-style report format (this is documented in Bracken's own output description), so `KrakenParser`'s existing Kraken2 report parser already handles it correctly - no new parsing code was needed. To make this verifiable rather than just asserted, I added a Bracken-style report fixture (`tests/conftest.py`) and a test (`tests/test_integration.py::test_kreport_to_mpa_parses_bracken_report`) proving the `mpa` step parses it identically to a Kraken2 report.

Metabuli is a different report format that `KrakenParser` does not actually parse, yet Metabuli supports Kraken2-like output format with additional parameters. I decided to remove all mentions of Metabuli from the manuscript and bibliography. The Summary now reads: "...reports generated by `Kraken2` and `Bracken`, which reports its re-estimated abundances in the same Kraken2-style report format."

### 2. Combined visualization
This claim had no corresponding code or test anywhere in the repository, so rather than build a feature to match an unreviewed manuscript claim, I removed the "Combined Visualizations" bullet from the Visualization list in `paper.md`. The visualization capabilities described now match exactly the three plot types the `kpplot` module implements and tests cover: stacked bar plots, streamgraphs, and clustermaps.

### 3. Performance and scalability
I removed "efficient," "scalable," "production-ready," and "large-scale" as self-descriptions of the software throughout the manuscript (including from the title, which now reads "KrakenParser: Post-processing of Kraken2-like Taxonomic Reports". I suppose, title change should be also approved by @crsl4), since no benchmark data backs these claims in this round. The one remaining use of "large-scale" in the Research Impact section describes the scale of the external study (Popov et al., 2025) that used the tool, not a performance claim about `KrakenParser` itself.

### 4. Command-line functionality on Windows
Confirmed and fixed. The root cause was non-ASCII characters (`➔`, `α`, `β`) in CLI help strings, docstrings, and log messages, which crash under Windows' default CP1252 console encoding. I replaced all of them with ASCII equivalents (`->`, "alpha", "beta") across `krakenparser/krakenparser.py`, `krakenparser/pipeline.py`, and `krakenparser/stats/diversity.py`, and updated the README examples to match. Thank you for noticing and flagging it!

I added a regression test, `tests/test_cli.py::test_source_has_no_non_ascii_cli_text`, that scans every source file under `krakenparser/` and fails if any non-ASCII character appears in our own text. This targets the actual crash — Typer/Rich's own decorative box-drawing help borders are handled safely by Rich's terminal-encoding detection and are deliberately excluded from this check, since flagging those would be a false positive unrelated to what actually failed on Windows. `tests/test_cli.py::test_help_exits_cleanly` additionally runs `--help` for the root command and every subcommand and asserts a clean exit code.

### 5. Automated tests
Added reference-value tests with hand-computed expected outputs for:
- Relative abundance (`tests/test_integration.py::test_relabund_reference_values`), including the `Other (<X%)` bucketing threshold.
- Bray-Curtis and Jaccard distances (`tests/test_integration.py::test_beta_div_reference_values`), using a fixed seed so rarefaction is deterministic.
- Shannon, Pielou, and Chao1 (`tests/test_units.py`) already existed pre-review and remain in place.

I also extended the integration tests to assert expected sample/taxa counts (matrix shape) rather than just "file exists and is non-empty," and added the Bracken fixture/test described in point 1.

### 6. User support and community contributions
Added `SUPPORT.md` at the repo root, covering where to ask usage questions, report bugs, and request features (all via GitHub Issues), plus a pointer to `CONTRIBUTING.md`. Linked it from the README's closing section. `CONTRIBUTING.md` now includes the actual development setup commands (`pip install -e ".[dev]"`, `pytest`, and how to run a single test file or keyword-filtered subset).

### 7. Manuscript text
Rewrote the Summary to state only verified capabilities, closely following your suggested wording: `KrakenParser` converts Kraken2-compatible taxonomic reports into rank-specific tables, calculates diversity metrics, and produces visualizations for bioinformaticians and microbiome researchers. I also fixed an unrelated text-encoding corruption in that paragraph.

### 8. Comparison with related software
Added a "Comparison with related software" section (`paper.md`) with a table comparing `KrakenParser` against `KrakenTools`, `Pavian`, and `spideog` across supported input, batch processing, diversity calculations, visualization, and mode of use. Every cell is grounded in the official documentation of the respective project, and no comparative claims of superiority are asserted — the table states what each tool does and leaves the comparison implicit.

### 9. References
- Corrected the Bray-Curtis citation to Bray, J.R. and Curtis, J.T. (1957), "An ordination of the upland forest communities of southern Wisconsin," *Ecological Monographs* 27(4):325-349.
- Replaced the `pandas` Zenodo `version=latest` entry with the canonical McKinney (2010) "Data Structures for Statistical Computing in Python" paper. Pandas' devs suggest 2 citing options on their website: https://pandas.pydata.org/about/citing.html. I decided to go with the 2nd option.
- Added citations for `spideog` (Manguy, 2020, software) and the MPA format (Segata et al., 2012, *Nature Methods*).
- Added `NumPy` (Harris et al., 2020, *Nature*) and `SciPy` (Virtanen et al., 2020, *Nature Methods*) as key scientific dependencies, since both are used directly in the diversity calculations.

All citation keys used in `paper.md` now resolve in `paper.bib` with no orphaned entries, and I verified this explicitly.

Once again thank you so much for the review! Thanks to your effort the tool became better!
