### 2026-09-02 #9694 @Ruslan-Kurmashev
Thank you very much @ilypopv for carefully addressing the review comments and for the substantial improvements made to both the software and its documentation.

One Functionality-related point remains before I can check this item. The manuscript currently states that rarefaction is performed before calculating the alpha-diversity indices and describes alpha diversity as including rarefaction. However, the implementation calculates Shannon, Pielou and Chao1 directly from the input count matrix, while the rarefaction depth is applied only during the calculation of Bray-Curtis and Jaccard beta-diversity matrices.

I confirmed this through an additional command-line test using rarefaction depths of 500 and 1000 with the same random seed. The resulting alpha_div.csv files were identical, whereas both beta-diversity matrices changed. This does not appear to be a software defect, but rather a remaining discrepancy between the manuscript and the implemented behavior. If the current implementation is intentional, please revise the comparison table and the Alpha Diversity description to clarify that alpha-diversity indices are calculated from the input count matrix, while rarefaction is applied only to beta-diversity calculations.

Thank you again for the thoughtful response and the considerable work invested in these improvements. Once this remaining statement has been corrected, I should be able to check the Functionality item.
