### 2026-09-02 #9694 @ilypopv
Hey again!

Thank you for double-checking the functionality and identifying this discrepancy!

The current behavior is indeed intentional: alpha-diversity indices are calculated directly from the input count matrix, whereas rarefaction is applied during the computation of beta-diversity matrices.

I have updated both the comparison table and the Statistical Analysis section in the manuscript to explicitly clarify that alpha diversity is computed on raw counts and that rarefaction is applied exclusively to beta-diversity calculations.

These changes were made in the latest [Pull Request](https://github.com/PopovIILab/KrakenParser/pull/27).

Thank you again for your constructive review!
