### 2026-09-02 #9694 @Ruslan-Kurmashev
Thank you @ilypopv and @crsl4 for the opportunity to review this submission. I greatly appreciate the author's prompt, thorough and constructive responses, as well as the improvements made to the software, documentation, automated tests and manuscript throughout the review process.
I have reviewed the latest revisions and confirm that all of my comments have been satisfactorily addressed. I have no further concerns for revision.
From my perspective as a reviewer, the submission is now ready to proceed to the next stage of the JOSS publication process. Thank you again for your careful work and professional engagement during the review. Good Luck !
